pyproject.toml
[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "timeslotfinder"
version = "0.1.0"
description = "Find available meeting slots using Microsoft Graph API"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "anyio==4.4.0",
    "httpx==0.27.0",
    "keyring==25.2.1",
    "msal==1.28.0",
    "pendulum==3.0.0",
    "pydantic==2.7.1",
    "pydantic-settings==2.2.1",
    "PyYAML==6.0.1",
    "rich==13.7.1",
    "typer[all]==0.12.3",
]

[project.scripts]
timeslotfinder = "timeslotfinder.cli.app:app"

[project.optional-dependencies]
dev = [
    "black==24.4.2",
    "mypy==1.10.0",
    "pytest==8.1.1",
    "pytest-cov==5.0.0",
    "ruff==0.4.2",
    "types-PyYAML==6.0.12.20240311",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["timeslotfinder*"]

[tool.black]
line-length = 100
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py311"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

